// src/app/shared/models/models.ts
export interface searchDocotr{

  rowId:any;
  visitingPrice:any;
  priceCurrency:any;
  persoonName:any;
  personType:any;
  personTitle:any;
  personDegree:any;
  personJob:any;
  personId:any;
  personGender:any;
  centerName:any;
  centerDesc:any;
  centerLocationLon:any;
  centerLocationLat:any;
  centerAddress:any;
  centerLogo:any;
  centerType:any;
  centerPhone:any;
  centerMobile:any;
  contactPerson:any;
  centerStatus:any;
  countryName:any;
  countryKey:any;
  cityId:any;
  cityName:any;
  areaId:any;
  areaName:any;
  typeId:any;
  typeName:any;
  typeDesc:any;
  statusId:any;
  statusName:any;
  statusDesc:any;
  personMajor:any;
  personNationality:any;
  degreeName:any;
  degreeDesc:any;
  majorName:any;
  majorDesc:any;
  jobName:any;
  jobDesc:any;
  TimeSlots:TimeSlots[];
  rating?:any;
}

export interface TimeSlots{
  timeSlotId:any;
  timeKeyTo:any;
  timeKeyFrom:any;
  dateKey:any;
  personsInCenterId:any;
  TimeDim:TimeDim[];
}
export interface TimeDim{
  timeKey:any;
  timeHhmmss:any;
  timeHh:any;
  timeMm:any;
  timeSs:any;
  timeAmPm:any;
}
// ...
export interface blogs {
  img1: string;
  img2: string;
  heading5: string;
  count1: string;
  count2: string;
  date: number | string;
  heading4: string;
  name: string;
  heading3: string;
  paragraph: string;
  msg: string;
  rating?:any;
}